requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
sqlparse==0.5.3
numpy>=1.19.0

Pillow>=9.0.0